# apps/agent/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/threads/{thread_id}/runs/stream": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Stream",
        "operationId": "stream_v1_threads__thread_id__runs_stream_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Thread Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stream SSE. Cada evento tem `event:` igual ao campo `event` do payload.",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/StreamEvent"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/threads/{thread_id}/runs/resume": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Resume",
        "operationId": "resume_v1_threads__thread_id__runs_resume_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Thread Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Stream SSE. Cada evento tem `event:` igual ao campo `event` do payload.",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/StreamEvent"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/threads/{thread_id}/runs/wait": {
      "post": {
        "tags": [
          "runs"
        ],
        "summary": "Wait",
        "description": "Execução síncrona, para consumidores que não querem lidar com SSE.",
        "operationId": "wait_v1_threads__thread_id__runs_wait_post",
        "parameters": [
          {
            "name": "thread_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Thread Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreadStateOut"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Conflict"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Gateway Timeout"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InterruptOut": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "value": {
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "id",
          "value"
        ],
        "title": "InterruptOut"
      },
      "MessageOut": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "type": {
            "enum": [
              "human",
              "ai",
              "tool",
              "system"
            ],
            "title": "Type",
            "type": "string"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "tool_calls": {
            "items": {
              "$ref": "#/components/schemas/ToolCallOut"
            },
            "title": "Tool Calls",
            "type": "array"
          },
          "tool_call_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Tool Call Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Name"
          }
        },
        "required": [
          "id",
          "type",
          "content",
          "tool_calls",
          "tool_call_id",
          "name"
        ],
        "title": "MessageOut",
        "type": "object"
      },
      "ResumeIn": {
        "properties": {
          "value": {
            "title": "Value"
          },
          "interrupt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interrupt Id",
            "description": "Necessário só quando há mais de um interrupt pendente."
          }
        },
        "type": "object",
        "required": [
          "value"
        ],
        "title": "ResumeIn"
      },
      "RunIn": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "RunIn"
      },
...
      "ThreadStateOut": {
        "properties": {
          "thread_id": {
            "type": "string",
            "format": "uuid",
            "title": "Thread Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/MessageOut"
            },
            "type": "array",
            "title": "Messages"
          },
          "next": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Next"
          },
          "interrupts": {
            "items": {
              "$ref": "#/components/schemas/InterruptOut"
            },
            "type": "array",
            "title": "Interrupts"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "messages",
          "next",
          "interrupts"
        ],
        "title": "ThreadStateOut"
      },
      "ToolCallOut": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "args": {
            "additionalProperties": true,
            "title": "Args",
            "type": "object"
          }
        },
        "required": [
          "id",
          "name",
          "args"
        ],
        "title": "ToolCallOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "DoneEvent": {
        "properties": {
          "event": {
            "const": "done",
            "default": "done",
            "title": "Event",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "interrupted": {
            "title": "Interrupted",
            "type": "boolean"
          },
          "next": {
            "items": {
              "type": "string"
            },
            "title": "Next",
            "type": "array"
          }
        },
        "required": [
          "event",
          "run_id",
          "interrupted",
          "next"
        ],
        "title": "DoneEvent",
        "type": "object"
      },
      "ErrorEvent": {
        "properties": {
          "event": {
            "const": "error",
            "default": "error",
            "title": "Event",
            "type": "string"
          },
          "code": {
            "enum": [
              "thread_busy",
              "timeout",
              "internal"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "event",
          "code",
          "message"
        ],
        "title": "ErrorEvent",
        "type": "object"
      },
      "InterruptEvent": {
        "properties": {
          "event": {
            "const": "interrupt",
            "default": "interrupt",
            "title": "Event",
            "type": "string"
          },
          "interrupt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interrupt Id"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "event",
          "interrupt_id",
          "value"
        ],
        "title": "InterruptEvent",
        "type": "object"
      },
      "MessageEvent": {
        "properties": {
          "event": {
            "const": "message",
            "default": "message",
            "title": "Event",
            "type": "string"
          },
          "node": {
            "title": "Node",
            "type": "string"
          },
          "message": {
            "$ref": "#/components/schemas/MessageOut"
          }
        },
        "required": [
          "event",
          "node",
          "message"
        ],
        "title": "MessageEvent",
        "type": "object"
      },
      "RunStartedEvent": {
        "properties": {
          "event": {
            "const": "run_started",
            "default": "run_started",
            "title": "Event",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "event",
          "run_id",
          "thread_id"
        ],
        "title": "RunStartedEvent",
        "type": "object"
      },
      "TokenEvent": {
        "properties": {
          "event": {
            "const": "token",
            "default": "token",
            "title": "Event",
            "type": "string"
          },
          "message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Id"
          },
          "content": {
            "title": "Content",
            "type": "string"
          }
        },
        "required": [
          "event",
          "message_id",
          "content"
        ],
        "title": "TokenEvent",
        "type": "object"
      },
      "StreamEvent": {
        "discriminator": {
          "mapping": {
            "done": "#/components/schemas/DoneEvent",
            "error": "#/components/schemas/ErrorEvent",
            "interrupt": "#/components/schemas/InterruptEvent",
            "message": "#/components/schemas/MessageEvent",
            "run_started": "#/components/schemas/RunStartedEvent",
            "token": "#/components/schemas/TokenEvent"
          },
          "propertyName": "event"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/RunStartedEvent"
          },
          {
            "$ref": "#/components/schemas/TokenEvent"
          },
          {
            "$ref": "#/components/schemas/MessageEvent"
          },
          {
            "$ref": "#/components/schemas/InterruptEvent"
          },
          {
            "$ref": "#/components/schemas/DoneEvent"
          },
          {
            "$ref": "#/components/schemas/ErrorEvent"
          }
        ]
      }
    }
  }
}